requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
rich>=13.0.0
sse-starlette>=1.8.0
beeai-framework
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
psutil>=5.9.0
redis[hiredis]>=5.0.0